This PR introduces a spatial data structure used to find the nearest player for the purposes of despawning mobs.

Currently, every mob loops over every player per tick to determine if they are:
1. Eligible for despawning mobs (not in spectator mode)
2. The closest player to that mob

This is effectively an `O(n ^ 2)` operation with per-player mob spawning.

In this PR, we construct a K-d tree (k=3) with every despawn-eligible player every tick and use that to find the closest player to determine whether the mob should despawn. This is a very efficient data structure for nearest-neighbour searches (the implemented data structure is an adapted version from the book "Algorithms In A Nutshell").

I have also changed the despawnRanges config option to be represented as an EnumMap rather than a HashMap to make it more efficient.

I considered using the NearbyPlayers class, however it raises a couple complications that made me decide against using it in the end. It stores players by chunk, and the area you get is a square grid, rather than a disc which is what is desired here, so you potentially still have to filter through a lot of players. In addition, pillagers will only despawn if they are not in the middle of a raid but with a hard coded despawn range of 128. You'd have to redo the Mojang logic with `respawnWhenFarAway` to make this work well. Another factor is that the user can set the despawn ranges to whatever they want, and it's a lot of logic to make sure that this is properly captured in NearbyPlayers.

Ensuring all the behaviour was correct ended up being rather awkward; the code was too tightly coupled, and didn't even perform better, so I went with this approach instead. Let me know if you think it would ultimately be better to use NearbyPlayers though, I might be missing something.

#### Benchmarks:

I use /spreadplayers to spread 50 bots over a 1000 block radius. Then I allowed the server to run for a few minutes and took a spark report. [The latest 1.21.8 paper](https://spark.lucko.me/ww2X8iVXvW) took 1.8 ms/t in Mob.checkDespawn, and [this patch](https://spark.lucko.me/ww2X8iVXvW) takes 1.07 ms/t in Mob.checkDespawn (plus 0.01 ms/t for calculating the player K-d trees). The difference is likely to be larger with even more players online. In total, that's a 0.72 ms/t overall increase in performance, which is similar to my other performance optimisation, [TemptGoal](https://github.com/PaperMC/Paper/pull/11942).